Removed Tag entity
Tag is removed because assigning a physical tag to a participant is redundant. Instead we will be giving a participant a random tag, for which a random number (Integer) will be generated. Relation participant-tag will still present in the system and will be used to connect a participant to the route